+++ app/product/biz/model/product.go
@@ -29,7 +29,7 @@ type Product struct {
 	Name        string     `json:"name"`
 	Description string     `json:"description"`
 	Picture     string     `json:"picture"`
-	Price       float32    `json:"price"`
+	Price       uint64     `json:"price"`
 	Categories  []Category `json:"categories" gorm:"many2many:product_category"`
 }
